requirements.txt
OpenAI
click
pyyaml
